requirements.txt
asgiref==3.7.2
Django==4.2.5
django-admin-rangefilter==0.11.2
django-jalali-date==1.0.2
djangorestframework==3.14.0
gevent==23.9.1
greenlet==2.0.2
gunicorn==21.2.0
jdatetime==4.1.1
packaging==23.2
psycopg2==2.9.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
zope.event==5.0
zope.interface==6.0
